+++ src/components/Layout.js
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { Link } from 'gatsby';
+import Button from 'react-bootstrap/Button'
 import Navigation from '../components/Navigation';
 import logo from '../img/giannelli.tech banner.png';
-import '../css/main.css';
 
 const Layout = ({ children }) => {
   const [wid, setwid] = useState("0%");
@@ -18,9 +18,9 @@ const Layout = ({ children }) => {
   return (
     <>
       <header className="header">
-        <Link to='/'><img className="image" src={logo} alt="giannelli.tech" /></Link>
-        <button onClick={openNav}>Open Nav</button>
+        <Button variant="outline-secondary" onClick={openNav}>Open Nav</Button>
         <Navigation width={wid} closeNav={closeNav} />
+        <Link to='/'><img className="image" src={logo} alt="giannelli.tech" /></Link>
       </header>
       <main className="main">
         {children}