+++ src/pages/app.js
@@ -1,21 +1,24 @@
-import React from 'react'
-import Layout from '../components/Layout'
-import { Router, Link } from "@reach/router"
-import Default from '../components/Default'
-import Dashboard from '../components/StockQuotes/Dashboard'
-import DashboardDefault from '../components/StockQuotes/DashboardDefault'
-import Data from '../components/StockQuotes/Data'
-import ChuckNorrisJokeGenerator from '../components/ChuckNorrisJokeGenerator'
+import React from 'react';
+import Nav from 'react-bootstrap/Nav';
+import Layout from '../components/Layout';
+import { Router, Link } from '@reach/router';
+import Default from '../components/Default';
+import Dashboard from '../components/StockQuotes/Dashboard';
+import DashboardDefault from '../components/StockQuotes/DashboardDefault';
+import Data from '../components/StockQuotes/Data';
+import ChuckNorrisJokeGenerator from '../components/ChuckNorrisJokeGenerator';
+import 'bootstrap/dist/css/bootstrap.min.css';
+import '../css/main.css';
 
 const App = ({ location }) => {
   return (
     <Layout>
       <h1>Portfolio</h1>
-      <nav>
+      <Nav className="flex-column">
         <Link to='/app'>Projects List</Link>
         <Link to='/app/chuck-norris'>Chuck Norris Joke Generator</Link>
         <Link to='/app/stock-quotes'>Stock Quotes Search</Link>
-      </nav>
+      </Nav>
       <Router basepath="/app">
         <Default path="/" />
         <ChuckNorrisJokeGenerator path="/chuck-norris" />